requirements.txt
streamlit
pandas
numpy
asyncio
openpyxl
plotly_express
streamlit_option_menu
pyxlsb
xlsxwriter